pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "This application is designed to automate the creation and uploading of youtube videos to my channel and monetize my channel"
authors = [
    {name = "Shubs2002",email = "s.d.nakashe2002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    # Authentication and security dependencies
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.6,<1.0.0)",
    # Database - MongoDB dependencies
    "motor (>=3.3.0,<4.0.0)",
    "beanie (>=1.26.0,<2.0.0)",
    # Email dependencies
    "aiosmtplib (>=3.0.0,<4.0.0)",
    "jinja2 (>=3.1.0,<4.0.0)",
    # Validation and utilities
    "pydantic[email] (>=2.0.0,<3.0.0)",
    "redis (>=5.0.0,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.dev:main"
prod = "src.main:main"